SigilHive/database/requirements.txt
numpy
aiomysql
langchain
langchain-google-genai
python-dotenv
aiofiles
uvloop
confluent_kafka